Added a progress property test for quiescence detection
Initially I thought that it would be sufficient to have only the garbage actors take snapshots. However, there may be stale snapshots from their old inverse acquaintances! These old snapshots say that the garbage actors do have inconsistent refobs.

This shows that the garbage detected by our algorithm, as it stands, is not maximal after all. But if every actor takes a snapshot, it is maximal.