Made quiescence detection a bit less restrictive
Apparently I misunderstood my own paper! A refob is "unreleased" for quiescence detection purposes if it's in the `owners` or `created` set; we don't need to consider the `refs` set.

I have ideas how to fix the progress bug, so we only need snapshots from garbage actors to make progress. For the meantime, I added some tests in comments.